Feature 238 - Implementation of checks and validations - Event Notifications utils

+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_0/CallbackUrlsApiController.java
@@ -20,12 +20,11 @@
  */
 package com.forgerock.openbanking.aspsp.rs.store.api.openbanking.event.v3_0;
 
-import com.forgerock.openbanking.aspsp.rs.store.api.helper.EventsHelper;
 import com.forgerock.openbanking.aspsp.rs.store.repository.TppRepository;
 import com.forgerock.openbanking.aspsp.rs.store.repository.v3_0.events.CallbackUrlsRepository;
 import com.forgerock.openbanking.common.model.openbanking.v3_0.event.FRCallbackUrl1;
 import com.forgerock.openbanking.common.model.version.OBVersion;
-import com.forgerock.openbanking.exceptions.OBErrorResponseException;
+import com.forgerock.openbanking.common.services.openbanking.event.EventResponseUtilService;
 import com.forgerock.openbanking.model.Tpp;
 import io.swagger.annotations.ApiParam;
 import lombok.extern.slf4j.Slf4j;
@@ -35,7 +34,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
-import uk.org.openbanking.datamodel.event.*;
+import uk.org.openbanking.datamodel.event.OBCallbackUrl1;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
@@ -43,7 +42,6 @@ import java.security.Principal;
 import java.util.Collection;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Controller("CallbackUrlsApiV3.0")
 @Slf4j
@@ -51,10 +49,30 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
     private CallbackUrlsRepository callbackUrlsRepository;
     private TppRepository tppRepository;
+    private EventResponseUtilService eventResponseUtilService;
 
-    public CallbackUrlsApiController(CallbackUrlsRepository callbackUrlsRepository, TppRepository tppRepository) {
+    public CallbackUrlsApiController(CallbackUrlsRepository callbackUrlsRepository, TppRepository tppRepository, EventResponseUtilService eventResponseUtilService) {
         this.callbackUrlsRepository = callbackUrlsRepository;
         this.tppRepository = tppRepository;
+        this.eventResponseUtilService = eventResponseUtilService;
+    }
+
+    /**
+     * Provides a way to obtain the instance {@link EventResponseUtilService} injected in the parent instance
+     * @return the response util service used by instance
+     */
+    public EventResponseUtilService getEventResponseUtilService() {
+        return eventResponseUtilService;
+    }
+
+    /**
+     * Provides a way to override the properties of {@link EventResponseUtilService} for every child that extends this parent controller. <br />
+     * Set the version per controller instance. <br />
+     * Set is should have the meta section in the response.
+     */
+    protected void initialiseResponseUtil() {
+        this.eventResponseUtilService.setVersion(OBVersion.v3_0);
+        this.eventResponseUtilService.setShouldHaveMetaSection(false);
     }
 
     @Override
@@ -81,11 +99,15 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
             HttpServletRequest request,
 
             Principal principal
-    ) throws OBErrorResponseException {
+    ) {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         log.debug("Create new callback URL: {} for client: {}", obCallbackUrl1Param, clientId);
 
         // Check if callback URL already exists for TPP
         // https://openbanking.atlassian.net/wiki/spaces/DZ/pages/645367055/Event+Notification+API+Specification+-+v3.0#EventNotificationAPISpecification-v3.0-POST/callback-urls
+        // A TPP must only create a callback-url on one version
         final Optional<Tpp> isTpp = Optional.ofNullable(tppRepository.findByClientId(clientId));
 
         if (!isTpp.isPresent()) {
@@ -111,7 +133,7 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
         return ResponseEntity
                 .status(HttpStatus.CREATED)
-                .body(packageResponse(frCallbackUrl1));
+                .body(eventResponseUtilService.packageResponse(frCallbackUrl1));
     }
 
     @Override
@@ -132,10 +154,20 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
             Principal principal
     ) {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         return Optional.ofNullable(tppRepository.findByClientId(clientId))
                 .map(Tpp::getId)
                 .map(id -> callbackUrlsRepository.findByTppId(id))
-                .map(urls -> ResponseEntity.ok(packageResponse(urls)))
+                // A TPP must only create a callback-url on one version and then there must be only one and we take only the first
+                .map(urls -> {
+                    if (urls.isEmpty()) {
+                        log.warn("No CallbackURL found for client id '{}'", clientId);
+                        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+                    }
+                    return ResponseEntity.ok(eventResponseUtilService.packageResponse(urls));
+                })
                 .orElseGet(() ->
                         {
                             log.warn("No TPP found for client id '{}'", clientId);
@@ -171,14 +203,21 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
             HttpServletRequest request,
 
             Principal principal
-    ) throws OBErrorResponseException {
+    ) {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         final Optional<FRCallbackUrl1> byId = callbackUrlsRepository.findById(callbackUrlId);
 
         if (byId.isPresent()) {
             FRCallbackUrl1 frCallbackUrl1 = byId.get();
-            frCallbackUrl1.setObCallbackUrl(obCallbackUrl1Param);
-            callbackUrlsRepository.save(frCallbackUrl1);
-            return ResponseEntity.ok(packageResponse(frCallbackUrl1));
+            if(eventResponseUtilService.allowOperationByVersion(frCallbackUrl1.getObCallbackUrl().getData().getVersion())){
+                frCallbackUrl1.setObCallbackUrl(obCallbackUrl1Param);
+                callbackUrlsRepository.save(frCallbackUrl1);
+                return ResponseEntity.ok(eventResponseUtilService.packageResponse(frCallbackUrl1));
+            }else{
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("Callback URL: '" + callbackUrlId + "' can't be delete via an older API version.");
+            }
         } else {
             // Spec isn't clear on if we should
             // 1. Reject a PUT for a resource id that does not exist
@@ -208,42 +247,21 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
             HttpServletRequest request,
 
             Principal principal
-    ) throws OBErrorResponseException {
+    ) {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         final Optional<FRCallbackUrl1> byId = callbackUrlsRepository.findById(callbackUrlId);
         if (byId.isPresent()) {
-            log.debug("Deleting callback url: {}", byId.get());
-            callbackUrlsRepository.deleteById(callbackUrlId);
-            return ResponseEntity.noContent().build();
+            if(eventResponseUtilService.allowOperationByVersion(byId.get().obCallbackUrl.getData().getVersion())){
+                log.debug("Deleting callback url: {}", byId.get());
+                callbackUrlsRepository.deleteById(callbackUrlId);
+                return ResponseEntity.noContent().build();
+            }else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("Callback URL: '" + callbackUrlId + "' can't be delete via an older API version.");
+            }
         } else {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Callback URL: '" + callbackUrlId + "' can't be found");
         }
-    }
-
-    protected OBCallbackUrlsResponse1 packageResponse(final Collection<FRCallbackUrl1> frCallbackUrls) {
-        return new OBCallbackUrlsResponse1()
-                .data(new OBCallbackUrlsResponseData1()
-                        .callbackUrl(
-                                frCallbackUrls.stream()
-                                        .filter(EventsHelper.matchingVersion(OBVersion.v3_0))
-                                        .map(this::toOBCallbackUrlResponseData1)
-                                        .collect(Collectors.toList())
-                        )
-                );
-    }
-
-
-    protected OBCallbackUrlResponse1 packageResponse(FRCallbackUrl1 frCallbackUrl) {
-        return new OBCallbackUrlResponse1()
-                .data(
-                        toOBCallbackUrlResponseData1(frCallbackUrl)
-                );
-    }
-
-    protected OBCallbackUrlResponseData1 toOBCallbackUrlResponseData1(FRCallbackUrl1 frCallbackUrl) {
-        final OBCallbackUrlData1 data = frCallbackUrl.getObCallbackUrl().getData();
-        return new OBCallbackUrlResponseData1()
-                .callbackUrlId(frCallbackUrl.getId())
-                .url(data.getUrl())
-                .version(data.getVersion());
     }
 }
