+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_2/callbackurl/CallbackUrlsApiController.java
@@ -26,6 +26,7 @@ import com.forgerock.openbanking.aspsp.rs.store.repository.v3_0.events.CallbackU
 import com.forgerock.openbanking.common.conf.discovery.ResourceLinkService;
 import com.forgerock.openbanking.common.model.openbanking.v3_0.event.FRCallbackUrl1;
 import com.forgerock.openbanking.common.model.version.OBVersion;
+import com.forgerock.openbanking.common.services.openbanking.event.EventResponseUtilService;
 import com.forgerock.openbanking.exceptions.OBErrorResponseException;
 import com.forgerock.openbanking.model.Tpp;
 import com.forgerock.openbanking.model.error.OBRIErrorResponseCategory;
@@ -45,7 +46,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.security.Principal;
 import java.util.Collection;
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -57,11 +57,31 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
     private CallbackUrlsRepository callbackUrlsRepository;
     private TppRepository tppRepository;
     private final ResourceLinkService resourceLinkService;
+    private final EventResponseUtilService eventResponseUtilService;
 
-    public CallbackUrlsApiController(CallbackUrlsRepository callbackUrlsRepository, TppRepository tppRepository, ResourceLinkService resourceLinkService) {
+    public CallbackUrlsApiController(CallbackUrlsRepository callbackUrlsRepository, TppRepository tppRepository, ResourceLinkService resourceLinkService, EventResponseUtilService eventResponseUtilService) {
         this.callbackUrlsRepository = callbackUrlsRepository;
         this.tppRepository = tppRepository;
         this.resourceLinkService = resourceLinkService;
+        this.eventResponseUtilService = eventResponseUtilService;
+    }
+
+    /**
+     * Provides a way to obtain the instance {@link EventResponseUtilService} injected in the parent instance
+     * @return the response util service used by instance
+     */
+    public EventResponseUtilService getEventResponseUtilService() {
+        return eventResponseUtilService;
+    }
+
+    /**
+     * Provides a way to override the properties of {@link EventResponseUtilService} for every child that extends this parent controller. <br />
+     * Set the version per controller instance. <br />
+     * Set is should have the meta section in the response.
+     */
+    protected void initialiseResponseUtil() {
+        this.eventResponseUtilService.setVersion(OBVersion.v3_1_2);
+        this.eventResponseUtilService.setShouldHaveMetaSection(true);
     }
 
     @Override
@@ -86,6 +106,9 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         log.debug("Create new callback URL: {} for client: {}", obCallbackUrl1Param, clientId);
 
         // https://openbanking.atlassian.net/wiki/spaces/DZ/pages/645367055/Event+Notification+API+Specification+-+v3.0#EventNotificationAPISpecification-v3.0-POST/callback-urls
@@ -121,7 +144,7 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
         return ResponseEntity
                 .status(HttpStatus.CREATED)
-                .body(packageResponse(frCallbackUrl1));
+                .body(eventResponseUtilService.packageResponse(frCallbackUrl1));
     }
 
     @Override
@@ -139,10 +162,19 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         return Optional.ofNullable(tppRepository.findByClientId(clientId))
                 .map(Tpp::getId)
                 .map(id -> callbackUrlsRepository.findByTppId(id))
-                .map(urls -> ResponseEntity.ok(packageResponse(urls)))
+                .map(urls -> {
+                    if (urls.isEmpty()) {
+                        log.warn("No CallbackURL found for client id '{}'", clientId);
+                        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+                    }
+                    return ResponseEntity.ok(eventResponseUtilService.packageResponse(urls));
+                })
                 .orElseThrow(() -> new OBErrorResponseException(
                                 HttpStatus.NOT_FOUND,
                                 OBRIErrorResponseCategory.REQUEST_INVALID,
@@ -176,13 +208,20 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         final Optional<FRCallbackUrl1> byId = callbackUrlsRepository.findById(callbackUrlId);
 
         if (byId.isPresent()) {
             FRCallbackUrl1 frCallbackUrl1 = byId.get();
-            frCallbackUrl1.setObCallbackUrl(obCallbackUrl1Param);
-            callbackUrlsRepository.save(frCallbackUrl1);
-            return ResponseEntity.ok(packageResponse(frCallbackUrl1));
+            if(eventResponseUtilService.allowOperationByVersion(frCallbackUrl1.getObCallbackUrl().getData().getVersion())){
+                frCallbackUrl1.setObCallbackUrl(obCallbackUrl1Param);
+                callbackUrlsRepository.save(frCallbackUrl1);
+                return ResponseEntity.ok(eventResponseUtilService.packageResponse(frCallbackUrl1));
+            }else{
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("Callback URL: '" + callbackUrlId + "' can't be update via an older API version.");
+            }
         } else {
             // Spec isn't clear on if we should
             // 1. Reject a PUT for a resource id that does not exist
@@ -214,11 +253,18 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         final Optional<FRCallbackUrl1> byId = callbackUrlsRepository.findById(callbackUrlId);
         if (byId.isPresent()) {
-            log.debug("Deleting callback url: {}", byId.get());
-            callbackUrlsRepository.deleteById(callbackUrlId);
-            return ResponseEntity.noContent().build();
+            if(eventResponseUtilService.allowOperationByVersion(byId.get().obCallbackUrl.getData().getVersion())){
+                log.debug("Deleting callback url: {}", byId.get());
+                callbackUrlsRepository.deleteById(callbackUrlId);
+                return ResponseEntity.noContent().build();
+            }else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("Callback URL: '" + callbackUrlId + "' can't be delete via an older API version.");
+            }
         } else {
             throw new OBErrorResponseException(
                     HttpStatus.BAD_REQUEST,
@@ -226,42 +272,5 @@ public class CallbackUrlsApiController implements CallbackUrlsApi {
                     OBRIErrorType.CALLBACK_URL_NOT_FOUND.toOBError1(callbackUrlId)
             );
         }
-    }
-
-
-    protected OBCallbackUrlsResponse1 packageResponse(final Collection<FRCallbackUrl1> frCallbackUrls) {
-        return new OBCallbackUrlsResponse1()
-                .data(new OBCallbackUrlsResponseData1()
-                        .callbackUrl(
-                                frCallbackUrls.stream()
-                                        .filter(EventsHelper.matchingVersion(OBVersion.v3_1_2))
-                                        .map(this::toOBCallbackUrlResponseData1)
-                                        .collect(Collectors.toList())
-                        )
-                );
-    }
-
-    protected List<OBCallbackUrlResponseData1> callbackUrlFilteredByVersion(final Collection<FRCallbackUrl1> frCallbackUrls, final String version){
-        return frCallbackUrls.stream()
-                .filter(
-                        frCallbackUrl1 -> frCallbackUrl1.obCallbackUrl.getData().getVersion().equals(version)
-                )
-                .map(this::toOBCallbackUrlResponseData1)
-                .collect(Collectors.toList());
-    }
-
-    protected OBCallbackUrlResponse1 packageResponse(FRCallbackUrl1 frCallbackUrl) {
-        return new OBCallbackUrlResponse1()
-                .data(toOBCallbackUrlResponseData1(frCallbackUrl))
-                .meta(new Meta())
-                .links(resourceLinkService.toSelfLink(frCallbackUrl, discovery -> discovery.getV_3_1_2().getGetCallbackUrls()));
-    }
-
-    protected OBCallbackUrlResponseData1 toOBCallbackUrlResponseData1(FRCallbackUrl1 frCallbackUrl) {
-        final OBCallbackUrlData1 data = frCallbackUrl.getObCallbackUrl().getData();
-        return new OBCallbackUrlResponseData1()
-                .url(data.getUrl())
-                .callbackUrlId(frCallbackUrl.getId())
-                .version(data.getVersion());
     }
 }
