+++ forgerock-openbanking-uk-aspsp-rs/forgerock-openbanking-uk-aspsp-rs-mock-store/forgerock-openbanking-uk-aspsp-rs-mock-store-server/src/main/java/com/forgerock/openbanking/aspsp/rs/store/api/openbanking/event/v3_1_2/eventsubscription/EventSubscriptionApiController.java
@@ -24,7 +24,8 @@ import com.forgerock.openbanking.aspsp.rs.store.repository.TppRepository;
 import com.forgerock.openbanking.aspsp.rs.store.repository.v3_1_2.events.EventSubscriptionsRepository;
 import com.forgerock.openbanking.common.conf.discovery.ResourceLinkService;
 import com.forgerock.openbanking.common.model.openbanking.event.FREventSubscription1;
-import com.forgerock.openbanking.common.services.openbanking.event.EventValidationService;
+import com.forgerock.openbanking.common.model.version.OBVersion;
+import com.forgerock.openbanking.common.services.openbanking.event.EventResponseUtilService;
 import com.forgerock.openbanking.exceptions.OBErrorResponseException;
 import com.forgerock.openbanking.model.Tpp;
 import com.forgerock.openbanking.model.error.OBRIErrorResponseCategory;
@@ -54,11 +55,32 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
     private final EventSubscriptionsRepository eventSubscriptionsRepository;
     private final TppRepository tppRepository;
     private final ResourceLinkService resourceLinkService;
+    private final EventResponseUtilService eventResponseUtilService;
 
-    public EventSubscriptionApiController(EventSubscriptionsRepository eventSubscriptionsRepository, TppRepository tppRepository, ResourceLinkService resourceLinkService) {
+
+    public EventSubscriptionApiController(EventSubscriptionsRepository eventSubscriptionsRepository, TppRepository tppRepository, ResourceLinkService resourceLinkService, EventResponseUtilService eventResponseUtilService) {
         this.eventSubscriptionsRepository = eventSubscriptionsRepository;
         this.tppRepository = tppRepository;
         this.resourceLinkService = resourceLinkService;
+        this.eventResponseUtilService = eventResponseUtilService;
+    }
+
+    /**
+     * Provides a way to obtain the instance {@link EventResponseUtilService} injected in the parent instance
+     * @return the response util service used by instance
+     */
+    public EventResponseUtilService getEventResponseUtilService() {
+        return eventResponseUtilService;
+    }
+
+    /**
+     * Provides a way to override the properties of {@link EventResponseUtilService} for every child that extends this parent controller. <br />
+     * Set the version per controller instance. <br />
+     * Set is should have the meta section in the response.
+     */
+    protected void initialiseResponseUtil() {
+        this.eventResponseUtilService.setVersion(OBVersion.v3_1_2);
+        this.eventResponseUtilService.setShouldHaveMetaSection(true);
     }
 
     public ResponseEntity createEventSubscription(
@@ -82,6 +104,9 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         log.debug("Create new event subscriptions: {} for client: {}", obEventSubscription, clientId);
 
         // Get the TPP from the header
@@ -119,7 +144,7 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
                 .body(packageResponse(frEventSubscription1));
     }
 
-    public ResponseEntity<OBEventSubscriptionsResponse1> readEventSubscription(
+    public ResponseEntity readEventSubscription(
             @ApiParam(value = "An Authorisation Token as per https://tools.ietf.org/html/rfc6750", required = true)
             @RequestHeader(value = "Authorization", required = true) String authorization,
 
@@ -133,22 +158,30 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
         log.debug("Read event subscription for client: {}", clientId);
-
-        return Optional.ofNullable(tppRepository.findByClientId(clientId))
-                .map(Tpp::getId)
-                .map(eventSubscriptionsRepository::findByTppId)
-                .map(eventSubs -> ResponseEntity.status(HttpStatus.OK).body(packageResponse(eventSubs)))
-                .orElseThrow( () ->
-                            new OBErrorResponseException(
-                                HttpStatus.NOT_FOUND,
-                                OBRIErrorResponseCategory.REQUEST_INVALID,
-                                OBRIErrorType.TPP_NOT_FOUND.toOBError1(clientId)
-                            )
-                );
+        final Optional<Tpp> tppId = Optional.ofNullable(tppRepository.findByClientId(clientId));
+        if (!tppId.isEmpty()) {
+            // A TPP must not access a event-subscription on an older version, via the EventSubscriptionId for an event-subscription created in a newer version
+            Collection<FREventSubscription1> frEventSubscription1s = Optional.ofNullable(eventSubscriptionsRepository.findByTppId(tppId.get().id))
+                    .orElseGet(Collections::emptyList)
+                    .stream().filter(frEventSubs -> eventResponseUtilService.allowOperationByVersion(frEventSubs.obEventSubscription1.getData().getVersion())).collect(Collectors.toList());
+            if (!frEventSubscription1s.isEmpty()) {
+                return ResponseEntity.status(HttpStatus.OK).body(packageResponse(frEventSubscription1s));
+            } else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("The event subscription can't be read via an older API version.");
+            }
+        } else {
+            throw new OBErrorResponseException(
+                    HttpStatus.NOT_FOUND,
+                    OBRIErrorResponseCategory.REQUEST_INVALID,
+                    OBRIErrorType.TPP_NOT_FOUND.toOBError1(clientId)
+            );
+        }
     }
 
-    public ResponseEntity<OBEventSubscriptionResponse1> updateEventSubscription(
+    public ResponseEntity updateEventSubscription(
             @ApiParam(value = "EventSubscriptionId", required = true)
             @PathVariable("EventSubscriptionId") String eventSubscriptionId,
 
@@ -172,6 +205,8 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
 
         final OBEventSubscription1 updatedSubscription =
                 new OBEventSubscription1().data(new OBEventSubscription1Data()
@@ -182,11 +217,14 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
         final Optional<FREventSubscription1> byId = eventSubscriptionsRepository.findById(eventSubscriptionId);
         if (byId.isPresent()) {
             FREventSubscription1 existingEventSubscription = byId.get();
-            EventValidationService.checkEqualOrNewerVersion(existingEventSubscription.getObEventSubscription1(), updatedSubscription);
-
-            existingEventSubscription.setObEventSubscription1(updatedSubscription);
-            eventSubscriptionsRepository.save(existingEventSubscription);
-            return ResponseEntity.ok(packageResponse(existingEventSubscription));
+            // A TPP must not update a event-subscription on an older version, via the EventSubscriptionId for an event-subscription created in a newer version
+            if (eventResponseUtilService.allowOperationByVersion(existingEventSubscription.obEventSubscription1.getData().getVersion())) {
+                existingEventSubscription.setObEventSubscription1(updatedSubscription);
+                eventSubscriptionsRepository.save(existingEventSubscription);
+                return ResponseEntity.ok(packageResponse(existingEventSubscription));
+            } else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("The event subscription can't be update via an older API version.");
+            }
         } else {
             // PUT is only used for amending existing subscriptions
             throw new OBErrorResponseException(
@@ -214,11 +252,19 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
 
             Principal principal
     ) throws OBErrorResponseException {
+        // initialise response util with the version and properties
+        initialiseResponseUtil();
+
         final Optional<FREventSubscription1> byId = eventSubscriptionsRepository.findById(eventSubscriptionId);
         if (byId.isPresent()) {
-            log.debug("Deleting event subscriptions URL: {}", byId.get());
-            eventSubscriptionsRepository.deleteById(eventSubscriptionId);
-            return ResponseEntity.noContent().build();
+            // A TPP must not delete a event-subscription on an older version, via the EventSubscriptionId for an event-subscription created in a newer version
+            if (eventResponseUtilService.allowOperationByVersion(byId.get().obEventSubscription1.getData().getVersion())) {
+                log.debug("Deleting event subscriptions URL: {}", byId.get());
+                eventSubscriptionsRepository.deleteById(eventSubscriptionId);
+                return ResponseEntity.noContent().build();
+            } else {
+                return ResponseEntity.status(HttpStatus.CONFLICT).body("The event subscription can't be delete via an older API version.");
+            }
         } else {
             throw new OBErrorResponseException(
                     HttpStatus.BAD_REQUEST,
@@ -247,7 +293,7 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
                     .data(new OBEventSubscriptionsResponse1Data()
                             .eventSubscription(eventSubsByClient))
                     .meta(new Meta())
-                    .links(resourceLinkService.toSelfLink(discovery -> discovery.getV_3_1_2().getGetCallbackUrls()));
+                    .links(resourceLinkService.toSelfLink(discovery -> discovery.getVersion(eventResponseUtilService.getVersion()).getGetCallbackUrls()));
         }
     }
 
@@ -260,7 +306,7 @@ public class EventSubscriptionApiController implements EventSubscriptionApi {
                         .eventTypes(obEventSubs.getEventTypes())
                         .version(obEventSubs.getVersion())
                 )
-                .links(resourceLinkService.toSelfLink(frEventSubscription1, discovery -> discovery.getV_3_1_2().getGetCallbackUrls()))
+                .links(resourceLinkService.toSelfLink(frEventSubscription1, discovery -> discovery.getVersion(eventResponseUtilService.getVersion()).getGetCallbackUrls()))
                 .meta(new Meta());
     }
 }
